+++ hw6/api/src/main/java/ru/miroshka/market/api/dto/ProductDto.java
@@ -1,11 +1,13 @@
 package ru.miroshka.market.api.dto;
 
 
+import java.math.BigDecimal;
+
 public class ProductDto {
 
     private Long id;
     private String title;
-    private int cost;
+    private BigDecimal cost;
 
     public Long getId() {
         return id;
@@ -23,15 +25,15 @@ public class ProductDto {
         this.title = title;
     }
 
-    public int getCost() {
+    public BigDecimal getCost() {
         return cost;
     }
 
-    public void setCost(int cost) {
+    public void setCost(BigDecimal cost) {
         this.cost = cost;
     }
 
-    public ProductDto(Long id, String title, int cost) {
+    public ProductDto(Long id, String title, BigDecimal cost) {
         this.id = id;
         this.title = title;
         this.cost = cost;
