+++ hw6/core-service/src/main/resources/db/migration/V1__init.sql
@@ -2,7 +2,7 @@ create table products
 (
     id         bigserial,
     title      varchar(255) not null,
-    cost       int          not null,
+    cost       money        not null,
     created_at timestamp default current_timestamp,
     updated_at timestamp default current_timestamp,
     primary key (id)
@@ -12,7 +12,7 @@ create table orders
 (
     id          bigserial primary key,
     username    varchar(255) not null,
-    total_price int    not null,
+    total_price money   not null,
     address     varchar(255),
     phone       varchar(255),
     created_at  timestamp default current_timestamp,
@@ -27,7 +27,7 @@ create table order_items
     order_id          bigint not null references orders (id),
     quantity          int    not null,
     price_per_product int    not null,
-    price             int    not null,
+    price             money    not null,
     created_at        timestamp default current_timestamp,
     updated_at        timestamp default current_timestamp
